main functions implemented

+++ backend/internal/models/risk_event.go
@@ -2,6 +2,7 @@ package models
 
 import "time"
 
+// RiskEvent 风险事件
 type RiskEvent struct {
 	ID              int       `json:"id" db:"id"`
 	EventType       string    `json:"event_type" db:"event_type"`
@@ -9,6 +10,6 @@ type RiskEvent struct {
 	ContractAddress string    `json:"contract_address" db:"contract_address"`
 	TxHash          string    `json:"tx_hash" db:"tx_hash"`
 	Description     string    `json:"description" db:"description"`
-	Score           int       `json:"score" db:"score"`
+	Score           float64   `json:"score" db:"score"`
 	DetectedAt      time.Time `json:"detected_at" db:"detected_at"`
 }
